+++ Sources/BraintreeVenmo/BTVenmoClient.m
@@ -1,8 +1,4 @@
 #import "BTVenmoClient_Internal.h"
-#import "BTVenmoAccountNonce_Internal.h"
-#import "BTVenmoAppSwitchRequestURL.h"
-#import "BTVenmoAppSwitchReturnURL.h"
-#import "BTVenmoRequest_Internal.h"
 #import "UIKit/UIKit.h"
 
 // MARK: - Swift File Imports for Package Managers
@@ -23,6 +19,13 @@
 
 #else                                            // Carthage
 #import <BraintreeCore/BraintreeCore-Swift.h>
+#endif
+
+// MARK: - Temporary Swift Module Imports
+#if __has_include(<Braintree/BraintreeVenmo.h>) // CocoaPods
+#import <Braintree/Braintree-Swift.h>
+#else                                            // SPM and Carthage
+#import <BraintreeVenmo/BraintreeVenmo-Swift.h>
 #endif
 
 @interface BTVenmoClient () <BTAppContextSwitchClient>
@@ -173,14 +176,14 @@ static BTVenmoClient *appSwitchedClient;
                 return;
             }
             
-            NSURL *appSwitchURL = [BTVenmoAppSwitchRequestURL appSwitchURLForMerchantID:merchantProfileID
+            NSURL *appSwitchURL = [BTVenmoAppSwitchRedirectURL appSwitchURLForMerchantID:merchantProfileID
                                                                             accessToken:configuration.venmoAccessToken
                                                                         returnURLScheme:self.returnURLScheme
                                                                       bundleDisplayName:bundleDisplayName
                                                                             environment:configuration.venmoEnvironment
                                                                        paymentContextID:paymentContextID
                                                                                metadata:self.apiClient.metadata];
-            
+
             [self performAppSwitch:appSwitchURL shouldVault:venmoRequest.vault completion:completionBlock];
         }];
     }];
@@ -245,7 +248,7 @@ static BTVenmoClient *appSwitchedClient;
 }
 
 - (BOOL)isiOSAppAvailableForAppSwitch {
-    return [self.application canOpenURL:[BTVenmoAppSwitchRequestURL baseAppSwitchURL]];
+    return [self.application canOpenURL:[BTVenmoAppSwitchRedirectURL baseAppSwitchURL]];
 }
 
 #pragma mark - App switch return