+++ Sources/BraintreeVenmo/Public/BraintreeVenmo/BraintreeVenmo.h
@@ -8,10 +8,6 @@ FOUNDATION_EXPORT const unsigned char BraintreeVenmoVersionString[];
 
 #if __has_include(<Braintree/BraintreeVenmo.h>)
 #import <Braintree/BTVenmoClient.h>
-#import <Braintree/BTVenmoAccountNonce.h>
-#import <Braintree/BTVenmoRequest.h>
 #else
 #import <BraintreeVenmo/BTVenmoClient.h>
-#import <BraintreeVenmo/BTVenmoAccountNonce.h>
-#import <BraintreeVenmo/BTVenmoRequest.h>
 #endif