+++ UnitTests/BraintreeVenmoTests/BraintreeVenmoTests-Bridging-Header.h
@@ -1,6 +1,3 @@
 #import <BraintreeVenmo/BTVenmoClient_Internal.h>
-#import <BraintreeVenmo/BTVenmoAppSwitchReturnURL.h>
-#import <BraintreeVenmo/BTVenmoAppSwitchRequestURL.h>
-#import <BraintreeVenmo/BTVenmoAccountNonce_Internal.h>
-#import <BraintreeVenmo/BTVenmoRequest_Internal.h>
 #import <BraintreeCore/BraintreeCore-Swift.h>
+#import <BraintreeVenmo/BraintreeVenmo-Swift.h>